requirements.txt
bintrees==2.0.7
boto==2.48.0
boto3==1.4.4
chardet==3.1.0
certifi==2017.7.27.1
crc16==0.1.1
dateparser==0.7.0
Django==1.11.13
django-allauth==0.36.0
django-betterforms==1.1.4
django-bootstrap3==10.0.1
django-bootstrap-form==3.4
django-bower==5.2.0
django-storages==1.6.5
djangorestframework-jwt==1.11.0
djangorestframework==3.6.3
ed25519==1.4
gdax==1.0.6
idna==2.7
mnemonic==0.18
oauthlib==2.0.7
pbkdf2==1.3
python-binance==0.6.9
python-magic==0.4.13
python-openid==2.2.5
pytz==2018.4
regex==2018.2.21
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
SSEClient==0.0.19
stellar-base==0.1.8.1
toml==0.9.4
tzlocal==1.5.1
urllib3==1.22
